feat(course-manager): Bổ sung thống kê Comment và Doanh thu vào danh sách khóa học
Mục đích:
Cập nhật API `posted-course` để trả về các số liệu thống kê cần thiết cho từng khóa học, bao gồm tổng số bình luận và tổng doanh thu đã thanh toán.

Các thay đổi chi tiết:
1. Cập nhật `CourseRepository`:
   - Thêm phương thức `@Query` `countCommentsByCourseId` để đếm tổng số bình luận cho một khóa học cụ thể.
   - Thêm phương thức `@Query` `sumRevenueByCourseIdAndPaymentStatusTrue` để tính tổng số tiền (`price`) từ các giao dịch `RegisteredCourse` có trạng thái thanh toán là `true`.

2. Cập nhật `CourseManagerController` (hoặc Controller liên quan):
   - Inject `CourseRepository` để truy cập các phương thức thống kê mới.
   - Trong `postedCourses()`, lặp qua danh sách `Course`:
     - Gọi `countCommentsByCourseId` và set giá trị cho `courseDTO.numberOfComment`.
     - Gọi `sumRevenueByCourseIdAndPaymentStatusTrue` và set giá trị cho `courseDTO.revenue` (xử lý trường hợp giá trị trả về là `NULL` thành `0.0`).

3. Cập nhật `CourseManagerDTO` (Giả định):
   - Đảm bảo DTO có sẵn các trường `numberOfComment` (int) và `revenue` (Double/double) để nhận dữ liệu thống kê.

Thay đổi này giúp người quản lý có cái nhìn tổng quan về hiệu suất của từng khóa học ngay trên trang danh sách.

+++ DATN_BE/src/main/java/com/fpoly/dto/CourseManagerDTO.java
@@ -17,7 +17,7 @@ public class CourseManagerDTO {
 	private int status;// 0 => không công khai, 1 = công khai, 2 = không công khai
 	private Date createAt;
 	private int numberOfComment;// Số lượng comment
-	private float revenue;// Doanh thu
+	private double revenue;// Doanh thu
 	private float courseDuration;// Thời lượng
 
 	private List<CategoryDto> categories;
